backend/pyproject.toml
[project]
name = "trendforge-backend"
version = "0.1.0"
description = "Real-Time AI Marketing Intelligence & Ad Generation System"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    
    # Redis
    "redis[hiredis]>=5.0.0",
    
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    
    # AI / ML
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.1.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",
    
    # NLP
    "nltk>=3.8.0",
    "textblob>=0.18.0",
    
    # HTTP clients
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    
    # Social APIs
    "tweepy>=4.14.0",
    "praw>=7.7.0",
    
    # Task queue
    "arq>=0.26.0",
    
    # Observability
    "structlog>=24.1.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    
    # Utils
    "python-dotenv>=1.0.0",
    "orjson>=3.10.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "factory-boy>=3.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "ICN", "PIE", "T20", "RET", "SIM"]
ignore = ["S101", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
